requirements.txt
aiodns==3.0.0
aiosmtpd==1.4.4.post2
asgiref==3.7.2
atpublic==4.0
attrs==23.1.0
cffi==1.15.1
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.1.1
psycopg2-binary==2.9.6
pycares==4.3.0
pycparser==2.21
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
verify-email==2.4.3
